+++ setup.py
@@ -69,12 +69,9 @@ setup(
         'dropbox>=9,<10',
     ],
 
-    # To provide executable scripts, use entry points in preference to the
-    # "scripts" keyword. Entry points provide cross-platform support and allow
-    # pip to create the appropriate form of executable for the target platform.
     entry_points={
         'console_scripts': [
-            'git-remote-dropbox=git_remote_dropbox:main',
+            'git-remote-dropbox=git_remote_dropbox.cli.git_remote_dropbox:main',
         ],
     },
 )
